backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
yahooquery>=2.3.7
pydantic>=2.0.0
exa-py>=1.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
